# Finishing gRPC server and refactoring controller
* server is using the controller now
* server returns the calculated discount
* refactoring the controller to return just the discount
