+++ discount-calculator/server/server_interface.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 type Server interface {
 	Run() <-chan error
